# docs/openapi_current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MITA Finance API",
    "version": "2.2.0",
...
    "contact": {
      "name": "MITA Finance API Support",
      "email": "api-support@mita.finance",
      "url": "https://docs.mita.finance/support"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://mita.finance/terms"
  },
  "servers": [
    {
      "url": "https://api.mita.finance",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.mita.finance",
      "description": "Staging Server"  
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "List User Transactions",
        "description": "Retrieve user transactions with filtering and pagination.\n\n**Features:**\n- Advanced filtering by date, category, amount\n- Pagination for large datasets\n- Real-time categorization\n- Comprehensive transaction metadata",
        "tags": ["Transactions"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of transactions per page"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["food", "dining", "groceries", "transportation", "gas", "public_transport", "entertainment", "shopping", "clothing", "healthcare", "insurance", "utilities", "rent", "mortgage", "education", "childcare", "pets", "travel", "subscriptions", "gifts", "charity", "other"]
            },
            "description": "Filter by transaction category"
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter transactions from this date"
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter transactions to this date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "headers": {
              "X-Total-Count": {
                "schema": {
                  "type": "integer"
                },
                "description": "Total number of transactions"
              },
              "X-Page-Count": {
                "schema": {
                  "type": "integer"
                },
                "description": "Total number of pages"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationInfo"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create New Transaction",
        "description": "Create a new financial transaction with comprehensive validation.\n\n**Features:**\n- Amount validation with proper limits (0.01 - 100,000)\n- Category validation against predefined list\n- Automatic budget impact assessment\n- Real-time financial analysis\n- Standardized error handling",
        "tags": ["Transactions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateRequest"
              },
              "examples": {
                "grocery_transaction": {
                  "summary": "Grocery Purchase",
                  "value": {
                    "amount": 67.89,
                    "category": "groceries",
                    "description": "Weekly grocery shopping at SuperMart",
                    "date": "2024-01-15T14:30:00Z"
                  }
                },
                "dining_transaction": {
                  "summary": "Restaurant Meal",
                  "value": {
                    "amount": 45.50,
                    "category": "dining",
                    "description": "Lunch at Italian restaurant"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Transaction"
                    },
                    "budget_impact": {
                      "$ref": "#/components/schemas/BudgetImpact"
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "txn_123abc456def",
                    "amount": 67.89,
                    "category": "groceries",
                    "description": "Weekly grocery shopping at SuperMart",
                    "date": "2024-01-15T14:30:00Z",
                    "created_at": "2024-01-15T14:30:00Z"
                  },
                  "budget_impact": {
                    "category": "groceries",
                    "amount": 67.89,
                    "remaining_budget": 432.11,
                    "budget_exceeded": false
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/transactions/receipt": {
      "post": {
        "summary": "Process Receipt with OCR",
        "description": "Extract transaction data from receipt image using OCR processing.\\n\\n**Features:**\\n- AI-powered OCR with Google Vision API\\n- Async processing with task tracking\\n- Premium user enhanced accuracy\\n- Automatic transaction extraction",
        "tags": ["Transactions", "OCR Processing"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Receipt image file (JPEG, PNG)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Receipt processing started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "task_id": {
                          "type": "string",
                          "example": "task_123abc456"
                        },
                        "status": {
                          "type": "string",
                          "example": "processing"
                        },
                        "estimated_completion": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "message": {
                          "type": "string",
                          "example": "Receipt processing started. Use /tasks/{task_id} to check status."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication. Format: `Bearer <jwt_token>`\n\n**Security Features:**\n- 30 minute expiration for access tokens\n- 90 day expiration for refresh tokens\n- Automatic token refresh capability\n- Token revocation support\n- Rate limiting protection\n\n**Usage:**\n```\nAuthorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...\n```"
      }
    },
    "schemas": {
...
      "TransactionCreateRequest": {
        "type": "object",
        "required": ["amount", "category", "description"],
        "properties": {
          "amount": {
            "type": "number",
            "format": "float",
            "minimum": 0.01,
            "maximum": 100000.0,
            "example": 67.89,
            "description": "Transaction amount in user's currency"
          },
          "category": {
            "type": "string",
            "enum": ["food", "dining", "groceries", "transportation", "gas", "public_transport", "entertainment", "shopping", "clothing", "healthcare", "insurance", "utilities", "rent", "mortgage", "education", "childcare", "pets", "travel", "subscriptions", "gifts", "charity", "other"],
            "example": "groceries",
            "description": "Transaction category"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "example": "Weekly grocery shopping at SuperMart",
            "description": "Transaction description"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T14:30:00Z",
            "description": "Transaction date (defaults to current time if not provided)"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "txn_123abc456def"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": 67.89
          },
          "category": {
            "type": "string",
            "example": "groceries"
          },
          "description": {
            "type": "string",
            "example": "Weekly grocery shopping at SuperMart"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T14:30:00Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T14:30:00Z"
          }
        }
      },
      "BudgetImpact": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "example": "groceries"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": 67.89
          },
          "remaining_budget": {
            "type": "number",
            "format": "float",
            "example": 432.11
          },
          "budget_exceeded": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "total_pages": {
            "type": "integer",
            "minimum": 1,
            "example": 5
          },
          "total_items": {
            "type": "integer",
            "minimum": 0,
            "example": 87
          },
          "items_per_page": {
            "type": "integer",
            "minimum": 1,
            "example": 20
          },
          "has_next_page": {
            "type": "boolean",
            "example": true
          },
          "has_previous_page": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "StandardizedErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for error responses"
          },
          "error": {
            "type": "object",
            "required": ["code", "message", "error_id", "timestamp"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_2002",
                "description": "Machine-readable error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "example": "Invalid input data",
                "description": "Human-readable error message"
              },
              "error_id": {
                "type": "string",
                "example": "mita_507f1f77bcf8",
                "description": "Unique error identifier for support and debugging"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "example": "2024-01-15T10:30:00Z",
                "description": "When the error occurred"
              },
              "details": {
                "type": "object",
                "description": "Additional error context and validation details"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardizedErrorResponse"
            },
            "examples": {
              "missing_token": {
                "summary": "Missing Authorization Header",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_MISSING",
                    "message": "Authorization header is required",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z"
                  }
                }
              },
              "invalid_token": {
                "summary": "Invalid Token",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_INVALID",
                    "message": "Invalid or malformed token",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z"
                  }
                }
              },
              "expired_token": {
                "summary": "Expired Token",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_EXPIRED",
                    "message": "Access token has expired",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "details": {
                      "expires_at": "2024-01-15T09:30:00Z",
                      "refresh_required": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation Error - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardizedErrorResponse"
            },
            "examples": {
              "field_validation": {
                "summary": "Field Validation Errors",
                "value": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_INVALID_FORMAT",
                    "message": "Invalid input data",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "details": {
                      "validation_errors": [
                        {
                          "field": "email",
                          "message": "Invalid email format",
                          "value": "invalid-email"
                        },
                        {
                          "field": "amount",
                          "message": "Amount must be positive",
                          "value": -50.0
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate Limit Exceeded - Too many requests",
        "headers": {
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            },
            "description": "Maximum requests allowed in time window"
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Requests remaining in current window"
          },
          "X-RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Unix timestamp when rate limit resets"
          },
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait before making another request"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardizedErrorResponse"
            },
            "examples": {
              "login_rate_limit": {
                "summary": "Login Rate Limit",
                "value": {
                  "success": false,
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Too many login attempts. Please try again in 15 minutes.",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "details": {
                      "limit": 5,
                      "window": "15 minutes",
                      "reset_at": "2024-01-15T10:45:00Z"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}